Bugs fixed
I noticed that in the portfolio, couple of things were not appropriate. I saw that by inspecting in phone device that, the navbar was not appropriate and the content was overflowing in the web page, to fix that I used javascript and made a Hamburger menu. In addition, I noticed that there were some formatting errors in the code so I used the W3C for my css and HTML to fix it. I guess it gonna be this much for today.